+++ packages/storybook-addon/src/utils/with-module-federation.ts
@@ -1,7 +1,7 @@
 import { getModuleFederationConfig } from '@nx/module-federation/src/with-module-federation/webpack/utils';
 import { container, Configuration } from 'webpack';
 
-import { ModuleFederationConfig } from '@nx/webpack';
+import { ModuleFederationConfig } from '@nx/module-federation';
 
 const { ModuleFederationPlugin } = container;