+++ packages/webpack-bundler-runtime/src/installInitialConsumes.ts
@@ -1,10 +1,14 @@
+import { getUsedExports } from './getUsedExports';
 import {
   HandleInitialConsumesOptions,
   InstallInitialConsumesOptions,
 } from './types';
+import type { Shared } from '@module-federation/runtime/types';
 import { updateConsumeOptions } from './updateOptions';
+
 function handleInitialConsumes(options: HandleInitialConsumesOptions) {
-  const { moduleId, moduleToHandlerMapping, webpackRequire } = options;
+  const { moduleId, moduleToHandlerMapping, webpackRequire, asyncLoad } =
+    options;
 
   const federationInstance = webpackRequire.federation.instance;
   if (!federationInstance) {
@@ -13,8 +17,22 @@ function handleInitialConsumes(options: HandleInitialConsumesOptions) {
   const { shareKey, shareInfo } = moduleToHandlerMapping[moduleId];
 
   try {
+    const usedExports = getUsedExports(webpackRequire, shareKey);
+
+    const customShareInfo: Partial<Shared> = { ...shareInfo };
+    if (usedExports) {
+      customShareInfo.treeShaking = {
+        usedExports,
+        useIn: [federationInstance.options.name],
+      };
+    }
+    if (asyncLoad) {
+      return federationInstance.loadShare(shareKey, {
+        customShareInfo,
+      });
+    }
     return federationInstance.loadShareSync(shareKey, {
-      customShareInfo: shareInfo,
+      customShareInfo,
     });
   } catch (err) {
     console.error(
@@ -26,21 +44,40 @@ function handleInitialConsumes(options: HandleInitialConsumesOptions) {
 }
 
 export function installInitialConsumes(options: InstallInitialConsumesOptions) {
-  const { webpackRequire } = options;
-
   updateConsumeOptions(options);
-  const { initialConsumes, moduleToHandlerMapping, installedModules } = options;
 
+  const {
+    moduleToHandlerMapping,
+    webpackRequire,
+    installedModules,
+    initialConsumes,
+    asyncLoad,
+  } = options;
+
+  const factoryIdSets: Array<
+    [string | number, () => Promise<false | (() => unknown)> | (() => unknown)]
+  > = [];
   initialConsumes.forEach((id) => {
+    const factoryGetter = () =>
+      handleInitialConsumes({
+        moduleId: id,
+        moduleToHandlerMapping,
+        webpackRequire,
+        asyncLoad,
+      }) as Promise<false | (() => unknown)>;
+    factoryIdSets.push([id, factoryGetter]);
+  });
+
+  const setModule = (
+    id: string | number,
+    factoryGetter: () => () => unknown,
+  ) => {
     webpackRequire.m[id] = (module) => {
       // Handle scenario when module is used synchronously
       installedModules[id] = 0;
       delete webpackRequire.c[id];
-      const factory = handleInitialConsumes({
-        moduleId: id,
-        moduleToHandlerMapping,
-        webpackRequire,
-      });
+
+      const factory = factoryGetter();
       if (typeof factory !== 'function') {
         throw new Error(
           `Shared module is not available for eager consumption: ${id}`,
@@ -68,5 +105,17 @@ export function installInitialConsumes(options: InstallInitialConsumesOptions) {
       }
       module.exports = result;
     };
+  };
+
+  if (asyncLoad) {
+    return Promise.all(
+      factoryIdSets.map(async ([id, factoryGetter]) => {
+        const result = await factoryGetter();
+        setModule(id, () => result as () => unknown);
+      }),
+    );
+  }
+  factoryIdSets.forEach(([id, factoryGetter]) => {
+    setModule(id, factoryGetter as () => () => unknown);
   });
 }