+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @module-federation/rsbuild-plugin
 
+## 0.23.0
+
+### Minor Changes
+
+- 3aa3cf9: feat(rsbuild-plugin): support build node mf assets
+
+### Patch Changes
+
+- @module-federation/enhanced@0.23.0
+- @module-federation/node@2.7.28
+- @module-federation/sdk@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/enhanced@0.22.1
+- @module-federation/sdk@0.22.1
+- @module-federation/node@2.7.27
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/enhanced@0.22.0
+- @module-federation/sdk@0.22.0
+- @module-federation/node@2.7.26
+
 ## 0.21.6
 
 ### Patch Changes