+++ packages/webpack-bundler-runtime/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @module-federation/webpack-bundler-runtime
 
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/runtime@0.23.0
+- @module-federation/sdk@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/runtime@0.22.1
+- @module-federation/sdk@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/runtime@0.22.0
+- @module-federation/sdk@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes